RELEASE: v6.0 (#17)
* FIX: cli_argument - send signal if no argument was given

* FIX: cli_argument - respect first argument is program-call itself on check for no arguments

* ADD: command_line - new SIGNAL and handler for arguments -file and -path

* ADD: common_tools_string - new functions append_number, append_character, number_to_hex_string, append_hex_string

* UNFINISHED: uart0_driver - driver to use on unitteset and for raspberry-pi

* ADD: common_tools_number - new functions readu16_LSB / readu16_MSB from byte-array

* ADD: thread-interface - new modules to enable posix-thread usage on mac-os and linux systems

* ADD: trace-parser - new module to parse received traces to display it

* ADD: tracer-threads - new modules to receive, parse and display traces

* FIX: spi0_driver_unittest - file-naming in description

* ADD: unittset - new unittest for print_traces thread

* ADD: unittest for print_trace_object

* FIX: local_module_satus - added attribute __UNUSED__ to avoid compiler warnings

* ADD: common_tools_string - new function substring()

* CHANGE: trace_parser - using of common-tools-string, added some debug outputs

* FIX: thread_parse_trace_object - thread can now be terminated

* ADD: unittest for parse_trace_object

* FIX: UNITTEST - math_module / ir_remote_sony - added common-tools to makefile

* ADD: Unittest for Parsing Trace Object

* FIX: TRX-DRIVER-INTERFACE - changed number of bytes from u8 to u16

* FIX: TRX-DRIVER-INTERFACE - changed number of bytes form u8 to u6

* FIX: trace_parser - incorrect data parsing (u32 was casted to u16)

* FIX: trace_object - added define for TRACE_FOOTER

* FIX: unittest_signal_slot_interface - added return type to function, correct usage of pointer for signal-send

* ADD: unittest - new check IS_BETWEEN(val, max, min)

* FIX: unittest_msg_executer - usage of new check IS_BETWEEN for runtime depended aspects

* FIX: unittest_cfg_file_parser - added __UNUSED__ to unused variable to avoid compiler warnings

* ADD: make_unittests - shell-script to build all unittests

* FIX: thread_read_trace_object - matching of trace-header, correct set of trace-object-length, resepect pointer on set p_com driver

* ADD: unittest for thread_read_trace_object

* REMOVE - spi0_driver_atmega_128p.h - file is replaced by spi0_driver.h

* FIX: make_cpu_avr - added define for __avr__

* FIX: make_cpu_unittest - added define for __unittest__

* FIX: cfg_driver_interface - maxe device name length set to 96, added new struct for device-name, TRX_DRIVER_CONFIGURATION now has a seperate field for device-name if running on raspberry pi or as unittest

* ADD: command_line_arguments - added new signal for DEVICE_NAME

* FIX: spi0_driver_raspberry_pi - using new field for device-name of driver-configuration

* FIX: rpi_protocol_host - using new field for device-name of driver-configuration

* UNFINISHED: usart0_driver_raspberry_pi - new driver module for receiving traces on raspberry pi

* CHANGE: thread_read_trace_object - module now expects the trx-driver interface to be set from outside

* FIX: unittest_read_trace_object - set driver interface to module-under-test

* FIX: tracer - check for correct length of tracedata on sending array

* FIX: trace_parser - added define to give compiler warnings if traces still active

* FIX: thread_interface - added function to initialize thread object

* FIX: usart0_driver_raspberry_pi - incorrect function calls, calling init method to initialize thread

* FIX: thread_read_trace_object - added init-function to respect thread-interface, read single bytes to check for trace-hedaer, reduced thread-sleep-ditme from 50000 to 900 us

* FIX: thread_print_trace_object - trace-output file is only opened at beginning of thread and closed on leaving thread

* FIX: thread_parse_trace_object - reduces thread-sleep-time from 50000 to 900 us

* FIX: make_cpu_raspberry_pi - added -pthread to CFLAGS

* FIX: make_targets - tracer is called from rpi_control_release directory

* FIX: thread_interface - check for uninitialized thread-objects before starting it

* FIX: unittest for trace-threads - calling init method of thread-objects

* FIX: unittest for trace-threads - calling init method of thread-objects

* FIX: sp0_driver_raspberry_pi - replaced incorrect header-include, replaced incorrect function signature

* FIX: cfg_file_parser - SIGNAL_CFG_COMPLETE is send after timeout of 150ms after program start if no cfg-file was given

* ADD: unittest-cfg-file-parser - added test-case for signal-complete timeout

* ADD: unittest cfg-file-parser

* RELEASE: v6.0

+++ src/command_management/command_buffer_interface.h
@@ -14,7 +14,7 @@
 // --------------------------------------------------------------------------------
 
 typedef u8 (*CMD_BUFFER_INTERFACE_START_READ_CALLBACK)		(void);
-typedef u8 (*CMD_BUFFER_INTERFACE_BYTES_AVAILABLE_CALLBACK)	(void);
+typedef u16 (*CMD_BUFFER_INTERFACE_BYTES_AVAILABLE_CALLBACK)	(void);
 typedef u8 (*CMD_BUFFER_INTERFACE_GET_BYTE_CALLBACK) 		(void);
 typedef u16 (*CMD_BUFFER_INTERFACE_GET_WORD_CALLBACK)		(void);
 typedef u32 (*CMD_BUFFER_INTERFACE_GET_LONG_CALLBACK)		(void);
