+++ src/driver/communication/spi/spi0_driver.h
@@ -36,13 +36,13 @@ void spi0_driver_power_off(void);
 /* this function gives information if and how many bytes have been received
  * since the last time this function was called
  */
-u8 spi0_driver_bytes_available(void);
+u16 spi0_driver_bytes_available(void);
 
 /*
  * get at maximum n bytes from the trx's modules internal buffer (the sw-buffer)
  * returns the number of bytes that have been copied into the given buffer
  */
-u8 spi0_driver_get_N_bytes(u8 num_bytes, u8* p_buffer_to);
+u16 spi0_driver_get_N_bytes(u16 num_bytes, u8* p_buffer_to);
 
 /*
  * copys at maximum n bytes into the trx-modules' internal buffer
@@ -50,7 +50,7 @@ u8 spi0_driver_get_N_bytes(u8 num_bytes, u8* p_buffer_to);
  * this function does not start any operation, only the buffer is prepared.
  * returns the number of byts that have been copied into the given buffer
  */
-u8 spi0_driver_set_N_bytes(u8 num_bytes, const u8* p_buffer_from);
+u16 spi0_driver_set_N_bytes(u16 num_bytes, const u8* p_buffer_from);
 
 /*!
  *
@@ -77,7 +77,7 @@ void spi0_driver_start_rx(u16 num_of_rx_bytes);
  * @param num_bytes
  * @param timeout_ms
  */
-void spi0_driver_wait_for_rx(u8 num_bytes, u16 timeout_ms);
+void spi0_driver_wait_for_rx(u16 num_bytes, u16 timeout_ms);
 
 /*
  * stops any active rx-operation
@@ -96,7 +96,7 @@ void spi0_driver_start_tx(void);
  * @param num_bytes
  * @param timeout_ms
  */
-void spi0_driver_wait_for_tx(u8 num_bytes, u16 timeout_ms);
+void spi0_driver_wait_for_tx(u16 num_bytes, u16 timeout_ms);
 
 /*
  * stops any active tx-operation, internal buffer will left unchanged
