+++ src/driver/communication/usart/usart0_driver_atmega1284p.c
@@ -169,7 +169,7 @@ void usart0_driver_power_off(void) {
 	USART0_TX_BUFFER_clear_all();
 }
 
-u8 usart0_driver_bytes_available (void) {
+u16 usart0_driver_bytes_available (void) {
 
 	#if defined TRACES_ENABLED && defined LOCAL_USART_RX_TRACES
 	{
@@ -185,11 +185,11 @@ u8 usart0_driver_bytes_available (void) {
 
 
 
-u8 usart0_driver_get_N_bytes (u8 num_bytes, u8* p_buffer_to) {
+u16 usart0_driver_get_N_bytes (u16 num_bytes, u8* p_buffer_to) {
 
 	PASS();	// local_usart_driver_get_N_bytes()
 
-	u8 num_bytes_available = USART0_RX_BUFFER_bytes_available();
+	u16 num_bytes_available = USART0_RX_BUFFER_bytes_available();
 
 	if (num_bytes < num_bytes_available) {
 		num_bytes_available = num_bytes;
@@ -204,7 +204,7 @@ u8 usart0_driver_get_N_bytes (u8 num_bytes, u8* p_buffer_to) {
 	return num_bytes_available;
 }
 
-u8 usart0_driver_set_N_bytes (u8 num_bytes, const u8* p_buffer_from) {
+u16 usart0_driver_set_N_bytes (u16 num_bytes, const u8* p_buffer_from) {
 	if (num_bytes > USART0_TX_BUFFER_size()) {
 		num_bytes = USART0_TX_BUFFER_size();
 	}
@@ -232,7 +232,7 @@ void usart0_driver_start_rx (u16 num_of_rx_bytes) {
 	local_usart_status_set(LOCAL_USART_STATUS_RX_ACTIVE);
 }
 
-void usart0_driver_wait_for_rx(u8 num_bytes, u16 timeout_ms) {
+void usart0_driver_wait_for_rx(u16 num_bytes, u16 timeout_ms) {
 	(void) num_bytes;
 	(void) timeout_ms;
 }
@@ -269,7 +269,7 @@ void usart0_driver_start_tx (void) {
 	USART0_TX_BUFFER_stop_read();
 }
 
-void usart0_driver_wait_for_tx(u8 num_bytes, u16 timeout_ms) {
+void usart0_driver_wait_for_tx(u16 num_bytes, u16 timeout_ms) {
 	(void) num_bytes;
 	(void) timeout_ms;
 }
